feat：为播客创作工作流新增播放器视图、制作视图、设置视图和终端日志组件
• 实现播放器视图，用于音频播放和报告展示。

• 创建制作视图，用于管理制作阶段和日志。

• 开发设置视图，用于收集用户输入的播客主题。

• 引入终端日志组件，以macOS风格终端界面显示日志。

• 通过毛玻璃面板效果和响应式设计增强用户界面。
