+++ Co-creation-projects/JJason-DeepCastAgent/backend/env.example
@@ -33,4 +33,5 @@ NO_PROXY=chat.ecnu.edu.cn,api.longcat.chat,open.bigmodel.cn
 # 服务器配置
 HOST=0.0.0.0
 PORT=8000
-CORS_ORIGINS=http://localhost:5173,http://localhost:3000
+LOG_LEVEL=INFO
+CORS_ORIGINS=http://localhost:5173,http://localhost:5174,http://localhost:3000
